requirements.txt
# Core dependencies
pyyaml
scikit-learn
pandas>=1.3.0
easydict
matplotlib>=3.4.0
networkx>=2.6.0
numpy>=1.20.0
scipy>=1.7.0
pyemd>=0.5.1
requests
tqdm

# For laplace approximation (DIFF-UQ)
laplace-torch

# For ChemNet semantic embeddings (molecular FCD)
fcd_torch>=1.0.5